% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBOFun.R
\name{CalcLambda}
\alias{CalcLambda}
\title{Calculate lambda and other attributes of the combinatoric train and validation matrices}
\usage{
CalcLambda(Train, Val, eval.method = "ave", risk.free.rate = 0.02,
  rank.ties.method = "random", verbose = FALSE)
}
\arguments{
\item{Train}{list of matrices}

\item{Val}{list of matrices, same length as Train}

\item{eval.method}{method of evaluating the columns (strategies), default is 'ave', but support 'SR'}

\item{risk.free.rate}{risk free rate for Sharpe ratio calculation, default is 0.02.}
}
\value{
list of elements : lambda, n.star, w.bar, IS.best, OOS
}
\description{
Calculate the lambda as defined in step (g) on p.12 of Bailey et al (2015).
Lambda is the logit of the relative rank (w.bar) of the best training strategy
in the matching validation set J_bar. It is calculated as follow,
For each train set,
1) find the column in train set with the best performance (eg. highest Sharpe ratio),
2) find its corresponding validation performance in the validation set Val,
3) calculate the rank of this column in Val,
4) lambda is the logit of this Val rank
}
\details{
The default method of evaluation is 'ave', which averages the columns. The best column
is the one with the maximum average. This would be appropriate if the matrix contains returns over
a fixed period. The alternative is 'SR', or the Sharpe ratio.
If the data are returns over a time period, it is better to use Sharpe ratio ("SR").
}
\examples{
   M = matrix(rnorm(N*TT, mean=0.1, sd=1), ncol=N, nrow=TT)
   Ms = DivideMat(M, S)
   res <- TrainValSplit(Ms)
   res1 <- CalcLambda(res$Train, res$Val, eval.method="ave")
   (Lambda = res1$lambda)
}
\references{
Bailey, D. H., Borwein, J., Lopez de Prado, M., & Zhu, Q. J. (2016). \emph{The probability of backtest overfitting}.
\url{https://www.carma.newcastle.edu.au/jon/backtest2.pdf}

Lopez de Prado (2018), \emph{Advances in Financial Machine Learning}, John Wiley & Sons.
}
\seealso{
\code{\link[PBO:PBO]{PBO::PBO()}}
}
\author{
Horace W. Tso \href{mailto:horacetso@gmail.com}{horacetso@gmail.com}
}
