% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBOFun.R
\name{DivideMat}
\alias{DivideMat}
\title{Divide a matrix into N chunks of equal size}
\usage{
DivideMat(M, n)
}
\arguments{
\item{M}{matrix with number of rows divisible by n}

\item{n}{number of submatrices to break M into}
}
\value{
list of length n, where each element is a matrix.
}
\description{
Split matrix M into equal chunks, where each chunk has the
same number of columns and rows. For example, if the original matrix has
10 columns and 100 rows, and n=5, this function generates five matrices of
size 20 x 10. Function returns a list of length n.
}
\examples{
   N = 20 # no of strategies
   TT = 1000 # no of observations
   S = 20 # no of partitions
   M = matrix(rnorm(N*TT, mean=0.1, sd=1), ncol=N, nrow=TT)
   Ms = DivideMat(M, S)
   length(Ms)
}
\author{
Horace W. Tso \href{mailto:horacetso@gmail.com}{horacetso@gmail.com}
}
