% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBOFun.R
\name{PBO}
\alias{PBO}
\title{Calculate the Probability of Backtest Overfitting (PBO)}
\usage{
PBO(lambda)
}
\arguments{
\item{lambda}{numeric vector, which is the relative rank logit, or lambda from CalcLambda()}
}
\value{
numeric value, which is the probability of backtest overfitting
}
\description{
Calculate the Probability of Backtest Overfitting as in section 11.6
of Lopez de Prado's book. It is the area of the lambda distribution to the
left of 0. In other words, it is the cumulative probability that the out-of-sample
relative rank of a strategy is lower (or worse) than the in-sample rank.\preformatted{    PBO = Integral_{-Inf}^{0} p(lambda) dlambda
}
}
\examples{
   M = matrix(rnorm(N*TT, mean=0.1, sd=1), ncol=N, nrow=TT)
   Ms = DivideMat(M, S)
   res <- TrainValSplit(Ms)
   res1 <- CalcLambda(res$Train, res$Val, eval.method="ave")
   (pbo = PBO(res1$lambda))
}
\references{
Bailey, D. H., Borwein, J., Lopez de Prado, M., & Zhu, Q. J. (2016). \emph{The probability of backtest overfitting}.
\url{https://www.carma.newcastle.edu.au/jon/backtest2.pdf}

Lopez de Prado (2018), \emph{Advances in Financial Machine Learning}, John Wiley & Sons.
}
\seealso{
\code{\link[PBO:CalcLambda]{PBO::CalcLambda()}}
}
\author{
Horace W. Tso \href{mailto:horacetso@gmail.com}{horacetso@gmail.com}
}
