% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBOFun.R
\name{TrainValSplit}
\alias{TrainValSplit}
\title{Form the train/validation split by generating all combinations of the given list of matrices
taking half of them at a time.}
\usage{
TrainValSplit(Ms)
}
\arguments{
\item{Ms}{list of equal-size matrices}
}
\value{
list of two lists : Train, Val, where each is a list of length n/2 of
matrices from the given Ms. Train <==> J, Val <==> J_bar in Bailey et al.
}
\description{
This function splits the data into a train set and a validation set as many times
as the number of combinations of drawing n/2 from n partitions. The number of this combinations
is given by Bin(n, n/2), or binomial coefficient of n items taking n/2 at a time. Each combination
provides the indices to the list Ms, and vertical stacking of these individual submatrices forms
a training matrix. What is not selected in a combination index set are used as the validation set.
}
\details{
If the order of the rows in the original matrix is chronological, then each of training and
validation matrix respects this chronological order.

WARNING : The resulting lists grow exponentially with the number of partitions. For example,
when N=10, there are 252 splits, so the length of the returning list is 504. Doubling N, the
list would have 369,512 matrices.
}
\examples{
   N = 20 # no of strategies
   TT = 1000 # no of observations
   S = 20 # no of partitions
   M = matrix(rnorm(N*TT, mean=0.1, sd=1), ncol=N, nrow=TT)
   Ms = DivideMat(M, S)
   res <- TrainValSplit(Ms)
   length(res$Train)
   length(res$Val)
   head(res$Train[[1]])
}
\references{
Bailey, D. H., Borwein, J., Lopez de Prado, M., & Zhu, Q. J. (2016). \emph{The probability of backtest overfitting}.
\url{https://www.carma.newcastle.edu.au/jon/backtest2.pdf}

Lopez de Prado (2018), \emph{Advances in Financial Machine Learning}, John Wiley & Sons.
}
\seealso{
\code{\link[PBO:CalcLambda]{PBO::CalcLambda()}}
}
\author{
Horace W. Tso \href{mailto:horacetso@gmail.com}{horacetso@gmail.com}
}
